+++ custom_components/blossom_energy/const.py
@@ -1,4 +1,4 @@
 """Constants for the Blossom Energy integration."""
 
 DOMAIN = "blossom_energy"
-INTEGRATION_VERSION = "0.4.1"
+INTEGRATION_VERSION = "0.4.2"